Add data receiving, sending confirmation

+++ WaterController v2.0.0/DaemonRPi/libs/src/Configuration.h
@@ -12,7 +12,7 @@ using namespace std;
 
 struct settings{
     string ipAddress;
-    uint32_t port;
+    unsigned short port;
 };
